+++ include/interpreter/polish_notation/polish_list.hpp
@@ -3,7 +3,6 @@
 #include <memory>
 #include <string>
 #include <vector>
-#include <iostream>
 #include "common/lexer_deque.hpp"
 #include "interpreter/polish_notation/polish.hpp"
 #include "exceptions/invalid_function_arg_exception.hpp"
@@ -23,7 +22,7 @@ class PolishList: public PolishFunction {
                                         std::shared_ptr<InterpreterContext>& context,
                                     const size_t fp_size) {
         std::vector<std::shared_ptr<SymObject>> elements;
-        for (uint32_t i = 0; i < get_num_args(); ++i) {
+        for (int64_t i = 0; i < get_num_args(); ++i) {
             auto element = iterate_wrapped(cmd_list, context, fp_size);
             elements.push_back(element);
         }
